fixed: fixed images for witch hat, cape (need to rename cloak), gloves and boots. also added defense points until we can resolve tool/use scenario.
