+++ client/src/components/footer/styles.scss
@@ -2,7 +2,7 @@
   font-size: 18px;
   text-transform: none;
   color:      #ffffff;
-  text-shadow: 2px 2px 5px indigo;
+  text-shadow: 0px 0px 5px indigo;
   margin: 35px 0 25px;
 
   align-self: center;
