+++ client/src/data/items/armor/cape/index.js
@@ -3,6 +3,7 @@ import capeImg from './cape.png';
 const cape = {
   name: 'Cape',
   type: 'armor::cape',
+  defence: 5,
 //   protection: spiders, spiderWebs,
   image: capeImg,
   value: 200
