+++ client/src/data/items/armor/witch-boots/index.js
@@ -1,8 +1,9 @@
-import witchBootsImg from './witchBoots.jpg';
+import witchBootsImg from './witch_boots.png';
 
 const witchBoots = {
   name: 'Witch Boots',
   type: 'armor::boots',
+  defence: 4
 //   protection: spiders, spiderWebs,
   image: witchBootsImg,
   value: 120
