+++ client/src/data/items/armor/witch-gloves/index.js
@@ -1,8 +1,9 @@
-import witchGlovesImg from './gloves.jpg';
+import witchGlovesImg from './gloves.png';
 
 const witchGloves = {
   name: 'Witch Gloves',
   type: 'armor::gloves',
+  defence: 3
 //   protection: spiders, spiderWebs,
   image: witchGlovesImg,
   value: 200
