+++ client/src/data/items/armor/witch-hat/index.js
@@ -3,6 +3,7 @@ import witchHatImg from './witch_hat.png';
 const witchHat = {
   name: 'Witch Hat',
   type: 'armor::hat',
+  defence: 2,
 //   protection: spiders, spiderWebs,
   image: witchHatImg,
   value: 200
